changed file structure and added dynamic input data
Changed the structure to be a bit more organized and updated all files to run from main.py. Also made it so everytime the grabber runs it grabs the new inputdata.txt file from github to elimate updating the entire docker file everytime I need to add a link. Also allows for easier pull requests to be made to the input data file.